# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MetaSRS Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  teacher: "teacher",
  student: "student",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  teacher: "teacher",
  student: "student",
} as const;

export interface CreateUserBody {
  name: string;
  email: string;
  role: CreateUserBodyRole;
}

export interface Class {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  subject: string;
  teacherId: number;
  /** @nullable */
  teacherName?: string | null;
  enrollmentCount: number;
  deckCount: number;
  createdAt: string;
}

export interface CreateClassBody {
  name: string;
  description?: string;
  subject: string;
  teacherId: number;
}

export interface UpdateClassBody {
  name?: string;
  description?: string;
  subject?: string;
}

export interface Enrollment {
  id: number;
  classId: number;
  studentId: number;
  enrolledAt: string;
}

export interface EnrollStudentBody {
  studentId: number;
}

export interface StudentEnrollment {
  studentId: number;
  studentName: string;
  studentEmail: string;
  enrolledAt: string;
  totalReviews: number;
  /** @nullable */
  averageRetention?: number | null;
  /** @nullable */
  lastReviewedAt?: string | null;
}

export interface Deck {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  classId?: number | null;
  teacherId: number;
  cardCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDeckBody {
  name: string;
  description?: string;
  classId?: number;
  teacherId: number;
}

export interface UpdateDeckBody {
  name?: string;
  description?: string;
  classId?: number;
}

export interface Card {
  id: number;
  deckId: number;
  front: string;
  back: string;
  /** @nullable */
  hint?: string | null;
  tags: string[];
  createdAt: string;
}

export interface CreateCardBody {
  front: string;
  back: string;
  hint?: string;
  tags?: string[];
}

export interface UpdateCardBody {
  front?: string;
  back?: string;
  hint?: string;
  tags?: string[];
}

export interface Review {
  id: number;
  studentId: number;
  cardId: number;
  deckId: number;
  /** 1=Again, 2=Hard, 3=Good, 4=Easy */
  grade: number;
  recalled: boolean;
  elapsedDays: number;
  /** @nullable */
  stabilityBefore?: number | null;
  /** @nullable */
  stabilityAfter?: number | null;
  /** @nullable */
  difficultyAfter?: number | null;
  nextReviewAt: string;
  reviewedAt: string;
}

export interface CreateReviewBody {
  studentId: number;
  cardId: number;
  deckId: number;
  /** 1=Again, 2=Hard, 3=Good, 4=Easy */
  grade: number;
  elapsedDays: number;
}

export interface DueCard {
  cardId: number;
  deckId: number;
  front: string;
  back: string;
  /** @nullable */
  hint?: string | null;
  /** @nullable */
  stabilityDays?: number | null;
  /** @nullable */
  difficulty?: number | null;
  /** @nullable */
  predictedRetention?: number | null;
  reviewCount: number;
  /** @nullable */
  lastReviewedAt?: string | null;
  /** @nullable */
  nextReviewAt?: string | null;
  isNew: boolean;
}

export interface ClassSummary {
  classId: number;
  className: string;
  studentCount: number;
  /** @nullable */
  averageRetention?: number | null;
  atRiskCount: number;
  totalReviews: number;
}

export interface TeacherAnalytics {
  teacherId: number;
  totalClasses: number;
  totalStudents: number;
  totalDecks: number;
  totalCards: number;
  totalReviews: number;
  /** @nullable */
  averageClassRetention?: number | null;
  classBreakdown: ClassSummary[];
}

export interface RetentionDataPoint {
  date: string;
  retention: number;
  reviewCount: number;
}

export interface MasteryDistribution {
  new: number;
  learning: number;
  reviewing: number;
  mastered: number;
}

export interface StruggleCard {
  cardId: number;
  front: string;
  averageGrade: number;
  totalReviews: number;
  recallRate: number;
}

export interface ClassAnalytics {
  classId: number;
  className: string;
  studentCount: number;
  deckCount: number;
  totalReviews: number;
  /** @nullable */
  averageRetention?: number | null;
  retentionTrend: RetentionDataPoint[];
  cardMasteryDistribution: MasteryDistribution;
  topStruggleCards: StruggleCard[];
}

export interface DeckProgress {
  deckId: number;
  deckName: string;
  totalCards: number;
  mastered: number;
  learning: number;
  new: number;
  dueToday: number;
  /** @nullable */
  averageRetention?: number | null;
}

export interface StudentAnalytics {
  studentId: number;
  studentName: string;
  totalReviews: number;
  cardsLearning: number;
  cardsMastered: number;
  cardsNew: number;
  /** @nullable */
  averageRetention?: number | null;
  currentStreak: number;
  retentionTrend: RetentionDataPoint[];
  deckProgress: DeckProgress[];
}

export type AtRiskStudentRiskLevel =
  (typeof AtRiskStudentRiskLevel)[keyof typeof AtRiskStudentRiskLevel];

export const AtRiskStudentRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface AtRiskStudent {
  studentId: number;
  studentName: string;
  studentEmail: string;
  /** @nullable */
  averageRetention?: number | null;
  /** @nullable */
  daysSinceLastReview?: number | null;
  cardsOverdue: number;
  riskLevel: AtRiskStudentRiskLevel;
  riskReason: string;
}

export type ListClassesParams = {
  teacherId?: number;
};

export type ListDecksParams = {
  classId?: number;
  teacherId?: number;
};

export type GetDueCardsParams = {
  deckId?: number;
};

export type ListStudentReviewsParams = {
  deckId?: number;
  limit?: number;
};