# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MetaSRS Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /decks/{deckId}/cards:
    get:
      operationId: listCards
      tags: [cards]
      summary: List cards in a deck
      parameters:
        - name: deckId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
    post:
      operationId: createCard
      tags: [cards]
      summary: Create a card in a deck
      parameters:
        - name: deckId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardBody"
      responses:
        "201":
          description: Created card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}:
    get:
      operationId: getCard
      tags: [cards]
      summary: Get a card by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCard
      tags: [cards]
      summary: Update a card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCardBody"
      responses:
        "200":
          description: Updated card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
    delete:
      operationId: deleteCard
      tags: [cards]
      summary: Delete a card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Reviews ─────────────────────────────────────────────────────────────────
...
  /analytics/teacher/{teacherId}:
    get:
      operationId: getTeacherAnalytics
      tags: [analytics]
      summary: Get teacher-level analytics across all classes
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teacher analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherAnalytics"
  /analytics/class/{classId}:
    get:
      operationId: getClassAnalytics
      tags: [analytics]
      summary: Get analytics for a specific class
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassAnalytics"
  /analytics/student/{studentId}:
    get:
      operationId: getStudentAnalytics
      tags: [analytics]
      summary: Get analytics for a student
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentAnalytics"
  /analytics/class/{classId}/at-risk:
    get:
      operationId: getAtRiskStudents
      tags: [analytics]
      summary: Get students showing memory slippage or at risk
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: At-risk students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AtRiskStudent"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Users ─────────────────────────────────────────────────────────────────
...
    Card:
      type: object
      properties:
        id:
          type: integer
        deckId:
          type: integer
        front:
          type: string
        back:
          type: string
        hint:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - deckId
        - front
        - back
        - tags
        - createdAt
    CreateCardBody:
      type: object
      properties:
        front:
          type: string
        back:
          type: string
        hint:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - front
        - back
    UpdateCardBody:
      type: object
      properties:
        front:
          type: string
        back:
          type: string
        hint:
          type: string
        tags:
          type: array
          items:
            type: string

    # ─── Reviews ───────────────────────────────────────────────────────────────
...
    TeacherAnalytics:
      type: object
      properties:
        teacherId:
          type: integer
        totalClasses:
          type: integer
        totalStudents:
          type: integer
        totalDecks:
          type: integer
        totalCards:
          type: integer
        totalReviews:
          type: integer
        averageClassRetention:
          type: ["number", "null"]
        classBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/ClassSummary"
      required:
        - teacherId
        - totalClasses
        - totalStudents
        - totalDecks
        - totalCards
        - totalReviews
        - classBreakdown
    ClassSummary:
      type: object
      properties:
        classId:
          type: integer
        className:
          type: string
        studentCount:
          type: integer
        averageRetention:
          type: ["number", "null"]
        atRiskCount:
          type: integer
        totalReviews:
          type: integer
      required:
        - classId
        - className
        - studentCount
        - atRiskCount
        - totalReviews
    ClassAnalytics:
      type: object
      properties:
        classId:
          type: integer
        className:
          type: string
        studentCount:
          type: integer
        deckCount:
          type: integer
        totalReviews:
          type: integer
        averageRetention:
          type: ["number", "null"]
        retentionTrend:
          type: array
          items:
            $ref: "#/components/schemas/RetentionDataPoint"
        cardMasteryDistribution:
          $ref: "#/components/schemas/MasteryDistribution"
        topStruggleCards:
          type: array
          items:
            $ref: "#/components/schemas/StruggleCard"
      required:
        - classId
        - className
        - studentCount
        - deckCount
        - totalReviews
        - retentionTrend
        - cardMasteryDistribution
        - topStruggleCards
    StudentAnalytics:
      type: object
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        totalReviews:
          type: integer
        cardsLearning:
          type: integer
        cardsMastered:
          type: integer
        cardsNew:
          type: integer
        averageRetention:
          type: ["number", "null"]
        currentStreak:
          type: integer
        retentionTrend:
          type: array
          items:
            $ref: "#/components/schemas/RetentionDataPoint"
        deckProgress:
          type: array
          items:
            $ref: "#/components/schemas/DeckProgress"
      required:
        - studentId
        - studentName
        - totalReviews
        - cardsLearning
        - cardsMastered
        - cardsNew
        - currentStreak
        - retentionTrend
        - deckProgress
    AtRiskStudent:
      type: object
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        studentEmail:
          type: string
        averageRetention:
          type: ["number", "null"]
        daysSinceLastReview:
          type: ["integer", "null"]
        cardsOverdue:
          type: integer
        riskLevel:
          type: string
          enum: [low, medium, high]
        riskReason:
          type: string
      required:
        - studentId
        - studentName
        - studentEmail
        - cardsOverdue
        - riskLevel
        - riskReason
    RetentionDataPoint:
      type: object
      properties:
        date:
          type: string
          format: date
        retention:
          type: number
        reviewCount:
          type: integer
      required:
        - date
        - retention
        - reviewCount
    MasteryDistribution:
      type: object
      properties:
        new:
          type: integer
        learning:
          type: integer
        reviewing:
          type: integer
        mastered:
          type: integer
      required:
        - new
        - learning
        - reviewing
        - mastered
    StruggleCard:
      type: object
      properties:
        cardId:
          type: integer
        front:
          type: string
        averageGrade:
          type: number
        totalReviews:
          type: integer
        recallRate:
          type: number
      required:
        - cardId
        - front
        - averageGrade
        - totalReviews
        - recallRate
    DeckProgress:
      type: object
      properties:
        deckId:
          type: integer
        deckName:
          type: string
        totalCards:
          type: integer
        mastered:
          type: integer
        learning:
          type: integer
        new:
          type: integer
        dueToday:
          type: integer
        averageRetention:
          type: ["number", "null"]
      required:
        - deckId
        - deckName
        - totalCards
        - mastered
        - learning
        - new
        - dueToday