# lib/api-zod/src/generated/types/card.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MetaSRS Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Card {
  id: number;
  deckId: number;
  front: string;
  back: string;
  /** @nullable */
  hint?: string | null;
  tags: string[];
  createdAt: Date;
}